Выполняет задание 5.16. Модуляция
